feat: Implement user account creation and preferred email update functionality; enhance error handling and validation
