+++ shareddump/Models/Authentication/AuthUserEntity.cs
@@ -39,11 +39,6 @@ public class AuthUserEntity : ITableEntity
     public string Email { get; set; } = string.Empty;
 
     /// <summary>
-    /// User's preferred email address for notifications (independent of auth provider email)
-    /// </summary>
-    public string? PreferredEmail { get; set; }
-
-    /// <summary>
     /// User's display name
     /// </summary>
     public string Name { get; set; } = string.Empty;
@@ -107,11 +102,5 @@ public class AuthUserEntity : ITableEntity
         CreatedAt = DateTime.UtcNow;
         LastLoginAt = DateTime.UtcNow;
         IsActive = true;
-        
-        // Set PreferredEmail to the provided email if available
-        if (!string.IsNullOrEmpty(email))
-        {
-            PreferredEmail = email;
-        }
     }
 }
